Refactor localstack code.
