+++ src/localstack/http/sqs/routes.go
@@ -1,9 +1,10 @@
-package http
+package sqs
 
 import "github.com/labstack/echo/v4"
 
 const prefix = "/localstack/sqs"
 
+// SetUpRoutes はsqs関連のルーティングをセットアップします
 func SetUpRoutes(e *echo.Echo) {
 	r := e.Group(prefix)
